+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/mcp
 
+## 1.0.0-beta.34
+
+### Patch Changes
+
+- Updated dependencies [960ec8f]
+  - @ai-sdk/provider-utils@4.0.0-beta.48
+
 ## 1.0.0-beta.33
 
 ### Patch Changes